warn about ptrdiff_t passed as long on 2-4 machines

+++ lang/cem/cemcom.ansi/LLlex.c
@@ -327,7 +327,7 @@ garbage:
 		if (ch != EOI)
 			UnGetChar();
 		*tg++ = '\0';	/* mark the end of the identifier	*/
-		idef = ptok->tk_idf = idf_hashed(buf, tg - buf, hash);
+		idef = ptok->tk_idf = idf_hashed(buf, (int) (tg - buf), hash);
 		idef->id_file = ptok->tk_file;
 		idef->id_line = ptok->tk_line;
 #ifndef	NOPP
